Space Key: only show locale if the user selects more than one
No need to show the locale unless there are more to switch to by swiping
on longpress space key.

I expect if the user has only selected one locale, then there is no need
to remind him which one that is, even if it's not the same as the system
language. I'm such a user ;)

Signed-off-by: Jean-Pierre Rasquin <yank555.lu@gmail.com>